requirements.txt
absl-py==2.2.2
accelerate==1.10.1
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
archspec @ file:///home/conda/feedstock_root/build_artifacts/archspec_1737352602016/work
atheris==3.0.0
attrs==25.3.0
av==15.1.0
babel==2.18.0
-e git+ssh://git@github.com/Krishnaa-Vinod/bay-max.git@25d1f5917c2da2c1abbb72c23032af7751d66c65#egg=baymax
beautifulsoup4==4.13.4
bitsandbytes==0.47.0
blis==1.3.3
boltons @ file:///home/conda/feedstock_root/build_artifacts/boltons_1733827268945/work
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1725267488082/work
catalogue==2.0.10
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1739515848642/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1725560558132/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1735929714516/work
click==8.2.1
cloudpathlib==0.23.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
conda @ file:///home/conda/feedstock_root/build_artifacts/conda_1736947900115/work
conda-libmamba-solver @ file:///home/conda/feedstock_root/build_artifacts/conda-libmamba-solver_1727359833193/work/src
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1736345463896/work
conda_package_streaming @ file:///home/conda/feedstock_root/build_artifacts/conda-package-streaming_1741620732069/work
confection==0.1.5
contourpy==1.3.3
coverage==7.13.5
csvw==3.7.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.13
datasets==4.0.0
decorator==5.2.1
decord==0.6.0
deepspeed @ file:///scratch/kvinod/CRM/DeepSpeed
deprecation==2.1.0
diffusers==0.33.1
dill==0.3.8
distro @ file:///home/conda/feedstock_root/build_artifacts/distro_1734729835256/work
dlinfo==2.0.0
docopt==0.6.2
dropbox==12.0.2
einops==0.8.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
espeakng-loader==0.2.4
expecttest==0.3.0
facenet-pytorch==2.6.0
fastapi==0.135.1
ffmpeg-python==0.2.0
ffmpy==1.0.0
filelock==3.13.1
fonttools==4.62.1
frozendict @ file:///home/conda/feedstock_root/build_artifacts/frozendict_1728841327252/work
frozenlist==1.7.0
fsspec==2024.6.1
ftfy==6.3.1
future==1.0.0
googlesearch-python==1.3.0
gradio==6.9.0
gradio_client==2.3.0
groovy==0.1.2
grpcio==1.74.0
h11==0.16.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
h5py==3.14.0
hf-xet==1.1.9
hjson==3.1.0
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.34.4
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
iniconfig==2.3.0
isodate==0.7.2
Jinja2==3.1.4
joblib==1.5.1
jsonpatch @ file:///home/conda/feedstock_root/build_artifacts/jsonpatch_1733814567314/work
jsonpointer @ file:///home/conda/feedstock_root/build_artifacts/jsonpointer_1725302935093/work
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kaolin==0.1
kiwisolver==1.5.0
kokoro==0.9.4
lance-namespace==0.6.0
lance-namespace-urllib3-client==0.6.0
lancedb==0.29.2
language-tags==1.2.0
lazy-loader==0.5
libmambapy @ file:///home/conda/feedstock_root/build_artifacts/mamba-split_1735835715095/work/libmambapy
librt==0.8.1
loguru==0.7.3
lxml==5.4.0
lz4==4.4.5
mamba @ file:///home/conda/feedstock_root/build_artifacts/mamba-split_1735835715095/work/mamba
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.8
mcap==1.3.1
mdurl==0.1.2
menuinst @ file:///home/conda/feedstock_root/build_artifacts/menuinst_1731146972698/work
misaki==0.9.4
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.6.4
multiprocess==0.70.16
murmurhash==1.0.15
mypy==1.19.1
mypy_extensions==1.1.0
networkx==3.3
ninja==1.11.1.4
nltk==3.9.1
num2words==0.5.14
numpy==1.26.4
nvdiffrast @ git+https://github.com/NVlabs/nvdiffrast@729261dc64c4241ea36efda84fbf532cc8b425b8
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.570.86
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open_clip_torch==2.32.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orjson==3.11.7
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandas==2.3.2
parameterized==0.9.0
pathspec==1.0.4
peft==0.17.1
perfetto==0.16.0
phonemizer-fork==3.3.2
pillow==10.2.0
platformdirs==4.9.4
pluggy==1.6.0
ply==3.11
portalocker==3.2.0
preshed==3.0.12
prettytable==3.17.0
proglog==0.1.12
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
py-cpuinfo==9.0.0
py-feat==0.6.2
pyarrow==21.0.0
pybind11==3.0.2
pycosat @ file:///home/conda/feedstock_root/build_artifacts/pycosat_1732588402431/work
pydantic==2.11.4
pydantic-settings==2.13.1
pydantic_core==2.33.2
pydot==4.0.1
pydub==0.25.1
pygame==2.6.1
Pygments==2.19.1
pymeshlab @ git+https://github.com/cnr-isti-vclab/PyMeshLab.git@ead5c61c2f7d263281165fd7a41b4b0cd8179623
pyparsing==3.3.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.2
qwen-vl-utils==0.0.8
rdflib==7.6.0
referencing==0.37.0
regex==2024.11.6
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1733217035951/work
rfc3986==1.5.0
rich==14.0.0
rosbags==0.11.0
rouge_score==0.1.2
rpds-py==0.30.0
ruamel.yaml==0.19.1
ruff==0.15.6
sacrebleu==2.5.1
safehttpx==0.1.7
safetensors==0.5.3
scikit-image==0.26.0
scikit-learn==1.7.0
scipy==1.16.0
segments==2.4.0
semantic-version==2.10.0
sentence-transformers==4.1.0
sentencepiece==0.2.1
setuptools==75.8.2
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
sounddevice==0.5.5
soundfile==0.13.1
soupsieve==2.7
spacy==3.8.11
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
starlette==0.52.1
stone==3.3.1
symengine==0.14.1
sympy==1.13.3
tabulate==0.9.0
tdqm==0.0.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.3.0
thinc==8.3.10
threadpoolctl==3.6.0
tifffile==2026.3.3
tiktoken==0.11.0
timm==1.0.15
tokenizers==0.22.0
tomlkit==0.13.3
torch==2.2.2
torchaudio==2.7.0+cu118
torchvision==0.17.2
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
transformers @ git+https://github.com/huggingface/transformers.git@b0db5a02f39ebd2ccffd7f8eb77091fda61f9a1e
triton==3.3.0
trl==0.22.1
truststore @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_truststore_1739009763/work
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.2.0
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1734859416348/work
uvicorn==0.42.0
uvloop==0.22.1
wasabi==1.1.3
watchfiles==1.1.1
wcwidth==0.2.13
weasel==0.4.3
websockets==16.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.1.2
xformers==0.0.30
xxhash==3.5.0
yarl==1.20.1
zipp==3.21.0
zstandard==0.23.0
